requirements-cpu.txt
networkx
torch==2.0.1+cpu --index-url https://download.pytorch.org/whl/cpu
torchvision==0.15.2+cpu --index-url https://download.pytorch.org/whl/cpu
groq
torch-summary
einops
scikit-image
opencv-python
pyclipper
shapely
requests
cryptography
freetype-py
aiohttp
tqdm
deepL
ImageHash
kornia
huggingface_hub
transformers
py3langid==0.2.2
sentencepiece
editdistance
numpy==1.26.4
tensorboardX
websockets
protobuf<6.0.0,>=3.20.2  # `google-genai` dependancy requirement
ctranslate2
colorama
openai==1.63.0
tiktoken
httpx==0.27.2  # stop before blocking change in 0.28.0
open_clip_torch
safetensors
pandas
onnxruntime
timm
omegaconf
pytorch-lightning
python-dotenv
nest-asyncio
marshmallow
cython
aioshutil
aiofiles
arabic-reshaper
pyhyphen
langcodes
manga-ocr
langdetect
accelerate
uvicorn
fastapi
pydantic==2.5.0
python-multipart
google-genai
rich
regex
--extra-index-url https://frederik-uni.github.io/manga-image-translator-rust/python/wheels/simple/
rusty-manga-image-translator
python-bidi
# Install pydensecrf from source
# pydensecrf@https://github.com/lucasb-eyer/pydensecrf/archive/refs/heads/master.zip
